Reworked Huawei module with threading and queues
After numerous tests in work environment, I noticed that the main think slowing the program were bigger connection timeouts on some of the devices (mainly Huawei). The best solution I found for this was using threading and queue instead of loops. This greatly improved speed and scalability. Soon adding to IOS module, too.